+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 1.6.0 (Next)
+- Minimum Ruby version increased to 2.1.
+- Minimum Puppet version increased to 3.7.
+
 ### 1.5.1
 - Slight cleanup and optimization.
 - Fixed check for no spec directories during RSpec Puppet helper.